App lista para versión tablet
Se han modificado los diseños para que ahora la app también sea funcional en tablets.
Se ha añadido el nuevo valor de PVi al excel.

+++ app/src/main/java/es/upv/oximetro/Utilities.java
@@ -7,5 +7,5 @@ import java.util.Map;
 
 public class Utilities {
 
-    public static List<HashMap<String, Integer>> datosPulsioximetro = new ArrayList<HashMap<String, Integer>>();
+    public static List<HashMap<String, Double>> datosPulsioximetro = new ArrayList<HashMap<String, Double>>();
 }
